# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: First-party Cognirise website and mini CMS API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List CMS documents
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/DocumentKind"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DocumentStatus"
        - name: market
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
  /documents/{documentId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a draft document
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Permanently delete an eligible document
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        "204":
          description: Document deleted
        "409":
          $ref: "#/components/responses/Conflict"
  /documents/{documentId}/revisions:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      operationId: listDocumentRevisions
      tags: [documents]
      summary: List document revisions
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Paginated revisions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRevisionPage"
  /documents/{documentId}/revisions/{revisionId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
      - $ref: "#/components/parameters/RevisionId"
    get:
      operationId: getDocumentRevision
      tags: [documents]
      summary: Get a document revision
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRevision"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{documentId}/submit:
    post:
      operationId: submitDocument
      tags: [documents]
      summary: Submit a document for review
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewSubmissionInput"
      responses:
        "200":
          description: Document submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "409":
          $ref: "#/components/responses/Conflict"
  /documents/{documentId}/publish:
    post:
      operationId: publishDocument
      tags: [documents]
      summary: Publish or schedule a document
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationInput"
      responses:
        "200":
          description: Publication accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "409":
          $ref: "#/components/responses/Conflict"
  /documents/{documentId}/rollback:
    post:
      operationId: rollbackDocument
      tags: [documents]
      summary: Restore a prior revision as a new draft
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackInput"
      responses:
        "200":
          description: Restored draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "409":
          $ref: "#/components/responses/Conflict"
  /documents/{documentId}/archive:
    post:
      operationId: archiveDocument
      tags: [documents]
      summary: Archive a document
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveInput"
      responses:
        "200":
          description: Archived document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{documentId}/preview:
    get:
      operationId: previewDocument
      tags: [documents]
      summary: Get a time-limited preview
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Renderable preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPreview"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cognirise_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: number, multipleOf: 1, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: number, multipleOf: 1, minimum: 1, maximum: 100, default: 25 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 200 }
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string }
    RevisionId:
      name: revisionId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: object
          additionalProperties: true
...
    DocumentKind:
      type: string
      enum: [person, partner, platform, publication, case-study]
    DocumentStatus:
      type: string
      enum: [draft, in-review, approved, scheduled, published, archived]
    Document:
      type: object
      required: [id, kind, slug, title, status, content, markets, revisionNumber, createdAt, updatedAt]
      properties:
        id: { type: string }
        kind:
          $ref: "#/components/schemas/DocumentKind"
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        title: { type: string }
        summary: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/DocumentStatus"
        content:
          type: object
          additionalProperties: true
        seo:
          $ref: "#/components/schemas/SeoMetadata"
        mediaIds:
          type: array
          items: { type: string }
        markets:
          type: array
          items: { type: string }
        revisionNumber: { type: number, multipleOf: 1, minimum: 1 }
        currentRevisionId: { type: ["string", "null"] }
        publishedRevisionId: { type: ["string", "null"] }
        scheduledAt: { type: ["string", "null"], format: date-time }
        publishedAt: { type: ["string", "null"], format: date-time }
        createdBy: { type: string }
        updatedBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DocumentInput:
      type: object
      additionalProperties: false
      required: [kind, slug, title, content, markets]
      properties:
        kind:
          $ref: "#/components/schemas/DocumentKind"
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 160 }
        title: { type: string, minLength: 1, maxLength: 200 }
        summary: { type: string, maxLength: 500 }
        content:
          type: object
          additionalProperties: true
        seo:
          $ref: "#/components/schemas/SeoMetadataInput"
        mediaIds:
          type: array
          items: { type: string }
        markets:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
    DocumentUpdate:
      type: object
      additionalProperties: false
      required: [revisionNumber]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 160 }
        title: { type: string, minLength: 1, maxLength: 200 }
        summary: { type: ["string", "null"], maxLength: 500 }
        content:
          type: object
          additionalProperties: true
        seo:
          $ref: "#/components/schemas/SeoMetadataInput"
        mediaIds:
          type: array
          items: { type: string }
        markets:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
        revisionNumber:
          type: number
          multipleOf: 1
          minimum: 1
          description: Optimistic concurrency version.
    SeoMetadata:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        canonicalUrl: { type: ["string", "null"] }
        noIndex: { type: boolean }
        imageId: { type: ["string", "null"] }
    SeoMetadataInput:
      type: object
      additionalProperties: false
      required: [title, description]
      properties:
        title: { type: string, minLength: 1, maxLength: 70 }
        description: { type: string, minLength: 1, maxLength: 170 }
        canonicalUrl: { type: string }
        noIndex: { type: boolean, default: false }
        imageId: { type: string }
    DocumentRevision:
      type: object
      required: [id, documentId, number, snapshot, createdBy, createdAt]
      properties:
        id: { type: string }
        documentId: { type: string }
        number: { type: number, multipleOf: 1, minimum: 1 }
        snapshot:
          $ref: "#/components/schemas/DocumentSnapshot"
        note: { type: ["string", "null"] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    DocumentSnapshot:
      type: object
      required: [slug, title, content, markets]
      properties:
        slug: { type: string }
        title: { type: string }
        summary: { type: ["string", "null"] }
        content:
          type: object
          additionalProperties: true
        seo:
          $ref: "#/components/schemas/SeoMetadata"
        mediaIds:
          type: array
          items: { type: string }
        markets:
          type: array
          items: { type: string }
    ReviewSubmissionInput:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 1000 }
        reviewerIds:
          type: array
          uniqueItems: true
          items: { type: string }
    PublicationInput:
      type: object
      additionalProperties: false
      required: [revisionId]
      properties:
        revisionId: { type: string }
        publishAt: { type: ["string", "null"], format: date-time }
        note: { type: string, maxLength: 1000 }
    RollbackInput:
      type: object
      additionalProperties: false
      required: [revisionId]
      properties:
        revisionId: { type: string }
        note: { type: string, maxLength: 1000 }
    ArchiveInput:
      type: object
      additionalProperties: false
      properties:
        reason: { type: string, maxLength: 1000 }
    DocumentPreview:
      type: object
      required: [document, previewUrl, expiresAt]
      properties:
        document:
          $ref: "#/components/schemas/DocumentSnapshot"
        previewUrl: { type: string }
        expiresAt: { type: string, format: date-time }
    DocumentPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Document"
    DocumentRevisionPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/DocumentRevision"
...
    PageMetadata:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: number, multipleOf: 1, minimum: 1 }
        pageSize: { type: number, multipleOf: 1, minimum: 1 }
        total: { type: number, multipleOf: 1, minimum: 0 }
        totalPages: { type: number, multipleOf: 1, minimum: 0 }